Remove user prop from Header and children
The user prop was not being used. It was being piped all the way down to
the FavoritesLink component, but the FavoritesLink component was not
using it.

I found this while investigating what we are using the user state for in
the application.

Removing this unnused prop will make it easier to see what the user
state is used for.